refactor: util이었던 NicknameGenerator 롤백
