+++ src/main/java/com/saebyeok/saebyeok/dto/CommentCreateRequest.java
@@ -3,7 +3,6 @@ package com.saebyeok.saebyeok.dto;
 import com.saebyeok.saebyeok.domain.Article;
 import com.saebyeok.saebyeok.domain.Comment;
 import com.saebyeok.saebyeok.domain.Member;
-import com.saebyeok.saebyeok.util.NicknameGenerator;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -27,11 +26,11 @@ public class CommentCreateRequest {
 
     private Long parentId;
 
-    public Comment toComment(Member member, Article article, Comment parent) {
+    public Comment toComment(Member member, String nickname, Article article, Comment parent) {
         return builder().
                 content(this.content).
                 member(member).
-                nickname(NicknameGenerator.generate(member, article)).
+                nickname(nickname).
                 article(article).
                 parent(parent).
                 build();
